BASIC FUNCTIONALITY 3
Added some steam animation to the burger and a trash system for if u want to throw the burger out. Gonna be honest, I might have to overhaul this system because I don't really see how I can turn this into an actual game in the state it's in. Will think on it more later.