BASIC FUNCTIONALITY 1
Decided to revise the entire thing because I feel like this would work better. Each ingredient is given a value which when added up together will equal the total amount a certain plate wants. Will add UI elements that show what each order is later, but this is just the basic building blocks of the system.